+++ wpull/options.py
@@ -958,6 +958,13 @@ class AppArgumentParser(argparse.ArgumentParser):
             action='store_true',
             help=_('don’t follow to parent directories on URL path'),
         )
+        group.add_argument(
+            '--no-strong-robots',
+            dest='strong_robots',
+            action='store_false',
+            default=True,
+            help=_('don’t ignore filters for robots.txt'),
+        )
 
     def _add_phantomjs_args(self):
         group = self.add_argument_group(_('PhantomJS'))